Fixed JsExpressionInterpreter class, changed long type to int

+++ task3/src/main/java/by/radchuk/task3/model/Expression.java
@@ -6,14 +6,14 @@ import java.util.Collections;
 import java.util.List;
 
 public class Expression extends TextElement {
-    private long value;
-    public Expression(List<TextElement> children, final long number) {
+    private int value;
+    public Expression(List<TextElement> children, final int number) {
         super(TextElementType.EXPRESSION, children);
         value = number;
     }
 
     @Override
     public String toString() {
-        return Long.toString(value);
+        return Integer.toString(value);
     }
 }
